removed old photo functionality
