--- .storybook/utils/addDefaultChildPhoto.ts
@@ -1,10 +1,0 @@
-import { Child } from "../../src/app/child-dev-project/children/model/child";
-import { BehaviorSubject } from "rxjs";
-import { SafeUrl } from "@angular/platform-browser";
-
-export function addDefaultChildPhoto(child: Child) {
-  child.photo = {
-    path: "",
-    photo: new BehaviorSubject<SafeUrl>("assets/child.png"),
-  };
-}
