+++ src/app/child-dev-project/children/child-block/child-block-tooltip/child-block-tooltip.component.ts
@@ -32,9 +32,9 @@ export class ChildBlockTooltipComponent implements OnInit {
   constructor(private fileService: FileService) {}
 
   ngOnInit() {
-    if (this.entity.photo2) {
+    if (this.entity.photo) {
       this.fileService
-        .loadFile(this.entity, "photo2")
+        .loadFile(this.entity, "photo")
         .subscribe((res) => (this.imgPath = res));
     }
   }
