+++ src/app/child-dev-project/children/children.module.ts
@@ -16,8 +16,6 @@
  */
 
 import { NgModule } from "@angular/core";
-import { EntitySchemaService } from "../../core/entity/schema/entity-schema.service";
-import { PhotoDatatype } from "./child-photo-service/datatype-photo";
 import { ComponentRegistry } from "../../dynamic-components";
 import { childrenComponents } from "./children-components";
 import { Aser } from "./aser/model/aser";
@@ -36,11 +34,7 @@ export class ChildrenModule {
     ChildSchoolRelation,
   ];
 
-  constructor(
-    entitySchemaService: EntitySchemaService,
-    components: ComponentRegistry
-  ) {
-    entitySchemaService.registerSchemaDatatype(new PhotoDatatype());
+  constructor(components: ComponentRegistry) {
     components.addAll(childrenComponents);
   }
 }
