+++ src/app/child-dev-project/children/model/child.ts
@@ -19,10 +19,6 @@ import { Entity } from "../../../core/entity/model/entity";
 import { DatabaseEntity } from "../../../core/entity/database-entity.decorator";
 import { DatabaseField } from "../../../core/entity/database-field.decorator";
 import { ConfigurableEnumValue } from "../../../core/configurable-enum/configurable-enum.interface";
-import { Photo } from "../child-photo-service/photo";
-import { BehaviorSubject } from "rxjs";
-import { SafeUrl } from "@angular/platform-browser";
-import { ChildPhotoService } from "../child-photo-service/child-photo.service";
 import { DateWithAge } from "./dateWithAge";
 import { IconName } from "@fortawesome/fontawesome-svg-core";
 
@@ -112,26 +108,16 @@ export class Child extends Entity {
   schoolClass: string;
 
   @DatabaseField({
-    dataType: "photo",
-    defaultValue: "",
+    dataType: "file",
     label: $localize`:Label for the filename of a photo of a child:Photo Filename`,
+    editComponent: "EditPhoto",
   })
-  photo: Photo = {
-    path: "",
-    photo: new BehaviorSubject<SafeUrl>(
-      ChildPhotoService.getImageFromAssets(undefined)
-    ),
-  };
+  photo: string;
 
   @DatabaseField({
     dataType: "file",
     label: $localize`:Label for the filename of a photo of a child:Photo Filename`,
     editComponent: "NewPhoto",
-  })
-  photo2: string;
-
-  @DatabaseField({
-    label: $localize`:Label for the phone number of a child:Phone Number`,
   })
   phone: string;
