--- src/app/core/entity-components/entity-utils/dynamic-form-components/edit-photo/new-photo.component.html
@@ -1,13 +1,0 @@
-<input type="file" style="display: none" (change)="onFileSelected($event)" #fileUpload>
-<img [src]="imgPath" alt="profile photo" class="child-pic">
-<div>
-  <button *ngIf="formControl.value && formControl.enabled" type="button" mat-icon-button
-          (click)="delete()"
-          i18n-mattooltip="Tooltip remove file" matTooltip="Remove file">
-    <fa-icon icon="xmark"></fa-icon>
-  </button>
-  <button *ngIf="formControl.enabled" type="button" mat-icon-button (click)="fileUpload.click()"
-          i18n-matTooltip="Tooltip upload file button" matTooltip="Upload file">
-    <fa-icon icon="upload"></fa-icon>
-  </button>
-</div>
