--- src/app/core/entity-components/entity-utils/dynamic-form-components/edit-photo/new-photo.component.scss
@@ -1,12 +1,0 @@
-.child-pic {
-  width: 150px;
-  height: 150px;
-  border-radius: 50%;
-  object-fit: cover;
-}
-
-:host {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
